% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renumChunks.R
\name{renumChunks}
\alias{renumChunks}
\title{Renumber all the code chunks in an Sweave/Knitr file or an R Markdown file}
\usage{
renumChunks(fi, dummyRun = TRUE, backup = TRUE, knitr = TRUE,
  labelStub = NULL)
}
\arguments{
\item{fi}{an object of class \code{fileInfo}. See \code{\link{setFileNames}}.}

\item{dummyRun}{if \code{FALSE} then the changes will be committed to disk as well as being displayed on screen. It is a good idea to do a dummy run first.}

\item{backup}{if \code{TRUE} and \code{dummyRun} is \code{FALSE} then a backup file will be created with sequential numbering to stop over-writing existing backups.}

\item{knitr}{if \code{TRUE} then knitr style code chunks are expected otherwise R Markdown code chunks are expected.}

\item{labelStub}{used for specifying a bespoke label stub for each label. If it is not specified (default) then \code{RC-} and the handout number are used.}
}
\description{
Given all the chunks a sequential unique ID
}
