% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renumPlotFilenames.R
\name{renumPlotFilenames}
\alias{renumPlotFilenames}
\title{Renumber all the PDF file names in an Sweave/Knitr file}
\usage{
renumPlotFilenames(fi, dummyRun = TRUE, backup = TRUE)
}
\arguments{
\item{fi}{an object of class \code{fileInfo}. See \code{\link{setFileNames}}.}

\item{dummyRun}{if \code{FALSE} then the changes will be committed to disk as well as being displayed on screen. It is a good idea to do a dummy run first.}

\item{backup}{if \code{TRUE} and \code{dummyRun} is \code{FALSE} then a backup file will be created with sequential numbering to stop over-writing existing backups.}
}
\description{
Makes sure the file
}
