% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimSummary.R
\name{slimSummary}
\alias{slimSummary}
\title{To be documented}
\usage{
slimSummary(lm.obj, showCall = !showAOVsummary, showResiduals = FALSE,
  showSignif = FALSE, showAOVsummary = ifelse(class(lm.obj)[1] == "lm",
  TRUE, FALSE), showGLMsummary = !showAOVsummary, ...)
}
\description{
To be documented
}
