% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimSummary.R
\name{slimSummary}
\alias{slimSummary}
\title{To be documented}
\usage{
slimSummary(
  lm.obj,
  showCall = FALSE,
  showResiduals = FALSE,
  showSignif = FALSE,
  showAOVsummary = TRUE,
  ...
)
}
\description{
To be documented
}
