% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimPlot.R
\name{trimPlot}
\alias{trimPlot}
\title{To be documented}
\usage{
<<<<<<< HEAD
trimPlot(
  x,
  data = NULL,
  fileName,
  plotCommand = plot,
  x.lab = "",
  y.lab = "",
  axes = FALSE,
  mai = c(0.5, 0.5, 0.1, 0.1),
  mgpx = c(3, 0.3, 0),
  mgpy = c(3, 0.6, 0),
  axis.size.cex = 0.7,
  axis.lab.cex = 0.7,
  linex = 1.2,
  liney = 1.8,
  lasy = 1,
  fig.width = 10,
  fig.height = 7,
  addElements = list(),
  ...
)
=======
trimPlot(x, data = NULL, fileName, plotCommand = plot, x.lab = "",
  y.lab = "", axes = FALSE, mai = c(0.5, 0.5, 0.1, 0.1), mgpx = c(3,
  0.3, 0), mgpy = c(3, 0.6, 0), axis.size.cex = c(0.7, 0.7),
  axis.lab.cex = 0.7, linex = 1.2, liney = 1.8, lasy = 1,
  fig.width = 10, fig.height = 7, .at = NULL, .labels = NULL,
  addElements = list(), ...)
>>>>>>> 37dd5fb2e8d1220633f05c73b066284723256ad5
}
\description{
To be documented
}
